alex/workflow-builder: separate out states
