+++ templates/workflow-builder/src/components/WorkflowRegions.tsx
@@ -10,9 +10,9 @@ import {
 	useQuickReactor,
 	useValue,
 } from 'tldraw'
+import { executionState, startExecution, stopExecution } from '../execution/executionState'
 import { NodeShape } from '../nodes/NodeShapeUtil'
 import { getNodePortConnections } from '../nodes/nodePorts'
-import { executionState, startExecution, stopExecution } from '../state'
 import { PlayIcon } from './icons/PlayIcon'
 import { StopIcon } from './icons/StopIcon'
