+++ templates/workflow-builder/src/connection/insertNodeWithinConnection.tsx
@@ -1,8 +1,8 @@
 import { Box, createShapeId, Editor, TLShapeId } from 'tldraw'
+import { onCanvasComponentPickerState } from '../components/OnCanvasComponentPicker'
 import { DEFAULT_NODE_SPACING_PX, NODE_WIDTH_PX } from '../constants'
 import { getNodePortConnections, getNodePorts } from '../nodes/nodePorts'
 import { NodeShape } from '../nodes/NodeShapeUtil'
-import { onCanvasComponentPickerState } from '../state'
 import { createOrUpdateConnectionBinding, getConnectionBindings } from './ConnectionBindingUtil'
 import { ConnectionShape } from './ConnectionShapeUtil'
